enhance(testing): add common BytesConcatenation to execution_testing
