+++ packages/testing/src/execution_testing/__init__.py
@@ -5,6 +5,7 @@ from .base_types import (
     Account,
     Address,
     Bytes,
+    BytesConcatenation,
     CoerceBytes,
     Hash,
     Storage,
@@ -143,6 +144,7 @@ __all__ = (
     "BlockException",
     "Bytecode",
     "Bytes",
+    "BytesConcatenation",
     "CalldataCase",
     "Case",
     "ChainConfig",
