+++ packages/testing/src/execution_testing/base_types/__init__.py
@@ -6,6 +6,7 @@ from .base_types import (
     BLSPublicKey,
     BLSSignature,
     Bytes,
+    BytesConcatenation,
     CoerceBytes,
     FixedSizeBytes,
     ForkHash,
@@ -56,6 +57,7 @@ __all__ = (
     "BLSPublicKey",
     "BLSSignature",
     "Bytes",
+    "BytesConcatenation",
     "CamelModel",
     "CoerceBytes",
     "EmptyOmmersRoot",
